LPS-171715 search-experiences-web: Move default error constant to errorMessages
